+++ src/main/java/com/breakoutms/lfs/server/undertaker/postmortem/PostmortemRequestRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.breakoutms.lfs.server.undertaker.postmortem.model.PostmortemRequest;
 
-public interface PostmortemRequestRepository extends JpaRepository<PostmortemRequest, Long> {
+public interface PostmortemRequestRepository extends JpaRepository<PostmortemRequest, Integer> {
 
 }
